feat(client)!: Completely refactor client package
