+++ packages/client/src/interceptors.ts
@@ -7,7 +7,7 @@
 import { Next, Headers } from '@temporalio/workflow';
 import { temporal } from '@temporalio/proto';
 import { CompiledWorkflowOptions } from './workflow-options';
-import { TerminateWorkflowExecutionResponse } from './types';
+import { TerminateWorkflowExecutionResponse, WorkflowExecution } from './types';
 
 export { Next, Headers };
 
@@ -18,27 +18,27 @@ export interface WorkflowStartInput {
   /** Workflow arguments */
   readonly args: unknown[];
   readonly headers: Headers;
+  readonly options: CompiledWorkflowOptions;
 }
 
 /** Input for WorkflowClientCallsInterceptor.signal */
 export interface WorkflowSignalInput {
   readonly signalName: string;
   readonly args: unknown[];
-  readonly namespace: string;
-  readonly workflowExecution: temporal.api.common.v1.IWorkflowExecution;
+  readonly workflowExecution: WorkflowExecution;
 }
 
 /** Input for WorkflowClientCallsInterceptor.query */
 export interface WorkflowQueryInput {
   readonly queryType: string;
   readonly args: unknown[];
-  readonly namespace: string;
-  readonly workflowExecution: temporal.api.common.v1.IWorkflowExecution;
+  readonly workflowExecution: WorkflowExecution;
+  readonly queryRejectCondition?: temporal.api.enums.v1.QueryRejectCondition;
 }
 
 /** Input for WorkflowClientCallsInterceptor.signal */
 export interface WorkflowTerminateInput {
-  readonly workflowExecution: temporal.api.common.v1.IWorkflowExecution;
+  readonly workflowExecution: WorkflowExecution;
   readonly reason?: string;
   readonly details?: unknown[];
 }
@@ -46,25 +46,30 @@ export interface WorkflowTerminateInput {
  * Implement any of these methods to intercept WorkflowClient outbound calls
  */
 export interface WorkflowClientCallsInterceptor {
-  start?: (input: WorkflowStartInput, next: Next<WorkflowClientCallsInterceptor, 'start'>) => Promise<unknown>;
-  signal?: (input: WorkflowSignalInput, next: Next<WorkflowClientCallsInterceptor, 'signal'>) => Promise<void>;
-  query?: (input: WorkflowQueryInput, next: Next<WorkflowClientCallsInterceptor, 'query'>) => Promise<unknown>;
+  start?: (input: WorkflowStartInput, next: Next<this, 'start'>) => Promise<string /* runId */>;
+  signal?: (input: WorkflowSignalInput, next: Next<this, 'signal'>) => Promise<void>;
+  query?: (input: WorkflowQueryInput, next: Next<this, 'query'>) => Promise<unknown>;
   terminate?: (
     input: WorkflowTerminateInput,
-    next: Next<WorkflowClientCallsInterceptor, 'terminate'>
+    next: Next<this, 'terminate'>
   ) => Promise<TerminateWorkflowExecutionResponse>;
+}
+
+interface WorkflowClientCallsInterceptorFactoryInput {
+  workflowId: string;
+  runId?: string;
 }
 
 /**
  * A function that takes {@link CompiledWorkflowOptions} and returns an interceptor
  */
 export interface WorkflowClientCallsInterceptorFactory {
-  (options: CompiledWorkflowOptions): WorkflowClientCallsInterceptor;
+  (input: WorkflowClientCallsInterceptorFactoryInput): WorkflowClientCallsInterceptor;
 }
 
 /**
  * A mapping of interceptor type of a list of factory functions
  */
-export interface ConnectionInterceptors {
-  workflowClient?: WorkflowClientCallsInterceptorFactory[];
+export interface WorkflowClientInterceptors {
+  calls?: WorkflowClientCallsInterceptorFactory[];
 }
