+++ packages/client/src/types.ts
@@ -1,5 +1,6 @@
 import { temporal } from '@temporalio/proto';
 
+export type WorkflowExecution = temporal.api.common.v1.IWorkflowExecution;
 export type StartWorkflowExecutionRequest = temporal.api.workflowservice.v1.IStartWorkflowExecutionRequest;
 export type GetWorkflowExecutionHistoryRequest = temporal.api.workflowservice.v1.IGetWorkflowExecutionHistoryRequest;
 export type DescribeWorkflowExecutionResponse = temporal.api.workflowservice.v1.IDescribeWorkflowExecutionResponse;
