+++ packages/create-project/samples/client-mtls.ts
@@ -1,6 +1,6 @@
 // @@@SNIPSTART nodejs-mtls-client
 import fs from 'fs';
-import { Connection } from '@temporalio/client';
+import { Connection, WorkflowClient } from '@temporalio/client';
 import { Example } from '../interfaces/workflows';
 import { getEnv, Env } from './mtls-env';
 
@@ -33,9 +33,10 @@ async function run({
     },
   });
   await connection.untilReady();
+  const client = new WorkflowClient(connection.service, { namespace });
   // Create a typed client using the Example Workflow interface,
-  const workflow = connection.workflow<Example>('example', { namespace, taskQueue });
-  const result = await workflow.start('Temporal');
+  const workflow = client.stub<Example>('example', { taskQueue });
+  const result = await workflow.execute('Temporal');
   console.log(result); // Hello, Temporal!
 }
